Added form and moved list rendering to a seoarate file
