+++ src/App.js
@@ -1,30 +1,14 @@
 import React from 'react';
-
-let todoList = [
-  {
-    id: 1,
-    title: "Wake up"
-  },
-  {
-    id: 2,
-    title: "Take a shower"
-  },
-  {
-    id: 3,
-    title: "Walk the dog"
-  }
-];
+import TodoList from './TodoList'; 
+import AddTodoForm from './AddTodoForm.js';
 
 function App() {
   return(
-    <span>
+    <div>
       <h1>ToDo List</h1>
-      <ul>
-        {todoList.map( (item) =>
-          <li key={item.id}>{item.title}</li>
-        )}
-      </ul>
-    </span>
+      <AddTodoForm />
+      <TodoList />
+    </div>
   )
 };
